Use a custom, NAR-specific assembly descriptor
This assembly descriptor is designed to incorporate all of the
platform-specific build projects available in the current target/nar/
directory.

The result is a .jar file with the -multi-platform.jar suffix that
includes all of the native libraries ready for use with the
native-lib-loader aware NarSystem class.

To use it, the project needs to be built from identical revisions for all
platforms that need to be supported. Then, the

	target/nar/<artifactId>-<version>-<aol>-jni/

directories need to be accumulated on one platform and a final `mvn` run
will produce the desired uber .jar file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>